# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Logistics Enquiry Triage Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/enquiries/{enquiry_id}/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Start Analysis",
        "description": "Start or reuse an analysis for an enquiry (single-flight, Fix 5).\n\nUses pg_advisory_xact_lock(enquiry_id) for the whole critical section:\n1. Check for an in-flight analysis (pending/in_progress) with same input_hash.\n2. If !force, check for a reusable completed analysis with same input_hash.\n3. Otherwise, allocate a new attempt number and create a new analysis + job.",
        "operationId": "start_analysis_enquiries__enquiry_id__analyses_post",
        "parameters": [
          {
            "name": "enquiry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Enquiry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis",
        "description": "Get analysis status, result, safety verdicts, and review prompt.",
        "operationId": "get_analysis_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "analyses"
        ],
        "summary": "Patch Analysis",
        "description": "Edit analysis result fields (\u00a74.2, ADR-9).\n\nUses If-Match header for optimistic concurrency (version check).\nOnly allowed when analysis.status == 'pending_review'.\nBLOCK verdict analyses are locked (no edits) unless admin override.",
        "operationId": "patch_analysis_analyses__analysis_id__patch",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Patch Analysis Analyses  Analysis Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Job Status",
        "description": "Get job status (Fix 6).",
        "operationId": "get_job_status_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/enquiries/{enquiry_id}/claim": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Claim Enquiry",
        "description": "Soft-claim an enquiry (ADR-9). Claims are not hard DB locks \u2014\nthey expire after CLAIM_TTL_MINUTES and can be overridden.",
        "operationId": "claim_enquiry_enquiries__enquiry_id__claim_post",
        "parameters": [
          {
            "name": "enquiry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Enquiry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Claim Enquiry Enquiries  Enquiry Id  Claim Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/enquiries/{enquiry_id}/release": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Release Enquiry",
        "description": "Release a soft claim. Reviewers can only release their own claims;\nadmins can release any claim.",
        "operationId": "release_enquiry_enquiries__enquiry_id__release_post",
        "parameters": [
          {
            "name": "enquiry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Enquiry Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/resume-review": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Resume Review",
        "description": "Resume a paused analysis review (\u00a74.2, Fix 1, Fix 10).\n\nActions:\n  - approve: approve the analysis as-is\n  - edit_and_approve: apply edited_fields and approve\n  - reject: reject the analysis\n  - override_block: admin-only, override a BLOCK verdict\n\nValidates:\n  - expected_analysis_version matches current version\n  - override_block requires admin role\n  - typed_reason required for certain actions (reject, override_block, quarantine confirm)\n  - edited_fields required for edit_and_approve and override_block",
        "operationId": "resume_review_analyses__analysis_id__resume_review_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Resume Review Analyses  Analysis Id  Resume Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "AnalysisRequest",
        "description": "Request body for POST /enquiries/{id}/analyses."
      },
      "AnalysisResponse": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "attached": {
            "type": "boolean",
            "title": "Attached"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "analysis_id",
          "attached",
          "reused"
        ],
        "title": "AnalysisResponse",
        "description": "Response for POST /enquiries/{id}/analyses (202)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}